requirements.txt
streamlit>=1.35.0
requests>=2.31.0
plotly>=5.18.0
pandas>=2.0.0
matplotlib>=3.8.0
fpdf2>=2.7.0
